Convert get-stylesheet to ts
